Add recomended changes
