+++ src/index.js
@@ -1,2 +1,5 @@
 import './style.css';
 import './modules/api.js';
+import { displayScores } from './modules/displayscores.js';
+
+displayScores();
