+++ src/modules/api.js
@@ -1,13 +1,14 @@
+import { displayScores } from './displayscores.js';
+
 const form = document.querySelector('.form');
 const refreshBtn = document.getElementById('refresh-btn');
-const table = document.getElementById('recent-list');
 const inputName = document.getElementById('name');
 const inputScore = document.getElementById('score');
 
 const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/iNLPApgiyuuDUgI7yzuD/scores/';
 
 const uploadScore = async (user, score) => {
-  const response = await fetch(`${url}`, {
+  const response = await fetch(url, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -18,24 +19,11 @@ const uploadScore = async (user, score) => {
   return data;
 };
 
-const getScores = async () => {
-  const response = await fetch(`${url}`);
-  const data = await response.json();
-  return data;
-};
-
-const refreshScores = async () => {
-  const scores = await getScores();
-
-  table.innerHTML = '';
-
-  scores.result.forEach((gamer) => {
-    const li = document.createElement('li');
-    li.classList.add('recent-score');
-    li.innerHTML = `${gamer.user}: ${gamer.score}`;
-    table.appendChild(li);
-  });
-};
+// const getScores = async () => {
+//   const response = await fetch(url);
+//   const data = await response.json();
+//   return data;
+// };
 
 form.addEventListener('submit', async (event) => {
   event.preventDefault();
@@ -50,5 +38,5 @@ form.addEventListener('submit', async (event) => {
 
 refreshBtn.addEventListener('click', (event) => {
   event.preventDefault();
-  refreshScores();
+  displayScores();
 });
